third_party/WebKit/LayoutTests/platform/win/external/wpt/html/browsers/history/the-location-interface/location-protocol-setter-non-broken-weird-expected.txt

CONSOLE ERROR: Not allowed to load local resource: blank.html
CONSOLE WARNING: Subresource requests using legacy protocols (like `ftp:`) are deprecated, and will be blocked in M59, around June 2017. Please deliver web-accessible resources over modern protocols like HTTPS. See https://www.chromestatus.com/feature/5709390967472128 for more details.
This is a testharness.js-based test.
PASS Set location.protocol to x 
FAIL Set location.protocol to data Blocked a frame with origin "http://web-platform.test:8001" from accessing a cross-origin frame.
FAIL Set location.protocol to file Failed to set the 'protocol' property on 'Location': 'file://web-platform.test:8001/common/blank.html' is not a valid URL.
PASS Set location.protocol to ftp 
PASS Set location.protocol to gopher 
PASS Set location.protocol to http+x 
Harness: the test ran to completion.
